Refactor(java): Move components into a components package
