+++ src/main/java/com/synopsys/integration/jira/common/cloud/model/components/IssueSeachIncludedFieldsComponent.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.jira.common.cloud.model;
+package com.synopsys.integration.jira.common.cloud.model.components;
 
 import java.util.List;
