+++ src/main/java/com/synopsys/integration/jira/common/cloud/model/components/StatusDetailsComponent.java
@@ -20,22 +20,51 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.jira.common.cloud.model;
+package com.synopsys.integration.jira.common.cloud.model.components;
 
 import com.synopsys.integration.rest.component.IntRestResponse;
 
-public class IdComponent extends IntRestResponse {
+public class StatusDetailsComponent extends IntRestResponse {
+    private String self;
+    private String description;
+    private String iconUrl;
+    private String name;
     private String id;
+    private StatusCategory statusCategory;
 
-    public IdComponent() {
+    public StatusDetailsComponent() {
     }
 
-    public IdComponent(final String id) {
+    public StatusDetailsComponent(final String self, final String description, final String iconUrl, final String name, final String id, final StatusCategory statusCategory) {
+        this.self = self;
+        this.description = description;
+        this.iconUrl = iconUrl;
+        this.name = name;
         this.id = id;
+        this.statusCategory = statusCategory;
+    }
+
+    public String getSelf() {
+        return self;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public String getIconUrl() {
+        return iconUrl;
+    }
+
+    public String getName() {
+        return name;
     }
 
     public String getId() {
         return id;
     }
 
+    public StatusCategory getStatusCategory() {
+        return statusCategory;
+    }
 }
