+++ src/main/java/com/synopsys/integration/jira/common/cloud/model/response/IssueSearchResponseModel.java
@@ -24,8 +24,8 @@ package com.synopsys.integration.jira.common.cloud.model.response;
 
 import java.util.List;
 
-import com.synopsys.integration.jira.common.cloud.model.IssueComponent;
-import com.synopsys.integration.jira.common.cloud.model.SchemaComponent;
+import com.synopsys.integration.jira.common.cloud.model.components.IssueComponent;
+import com.synopsys.integration.jira.common.cloud.model.components.SchemaComponent;
 import com.synopsys.integration.jira.common.model.JiraPageResponseModel;
 
 public class IssueSearchResponseModel extends JiraPageResponseModel {
