+++ src/main/java/com/synopsys/integration/jira/common/cloud/model/response/PageOfProjectsResponseModel.java
@@ -25,7 +25,7 @@ package com.synopsys.integration.jira.common.cloud.model.response;
 import java.util.Collections;
 import java.util.List;
 
-import com.synopsys.integration.jira.common.cloud.model.ProjectComponent;
+import com.synopsys.integration.jira.common.cloud.model.components.ProjectComponent;
 import com.synopsys.integration.jira.common.model.JiraPageResponseModel;
 
 public class PageOfProjectsResponseModel extends JiraPageResponseModel {
